Added the ability to view hotspots from the Android client. Optimizations to come.

+++ src/com/example/wifil/Hotspot.java
@@ -4,11 +4,11 @@ public class Hotspot {
 	
 	private String SSID = null;
 	private String Mac = null;
-	private long lat = 0;
-	private long lon = 0;
-	private long radius;
+	private double lat = 0;
+	private double lon = 0;
+	private double radius;
 	
-	public Hotspot(String SSID, String Mac, long lat, long lon, long radius) {
+	public Hotspot(String SSID, String Mac, double lat, double lon, double radius) {
 		this.SSID = SSID;
 		this.Mac = Mac;
 		this.lat = lat;
@@ -18,13 +18,13 @@ public class Hotspot {
 	
 	public String getSSID() {return this.SSID;} 
 	public String getMac() {return this.Mac;} 
-	public long getLat() {return this.lat;} 
-	public long getLon() {return this.lon;} 
-	public long getRadius() {return this.radius;} 
+	public double getLat() {return this.lat;} 
+	public double getLon() {return this.lon;} 
+	public double getRadius() {return this.radius;} 
 	
 	public void setSSID(String SSID) {this.SSID = SSID;} 
 	public void setMac(String Mac) {this.Mac = Mac;} 
-	public void setLat(long lat) {this.lat = lat;} 
-	public void setLon(long lon) {this.lon = lon;} 
-	public void setRadius(long radius) {this.radius = radius;} 
+	public void setLat(double lat) {this.lat = lat;} 
+	public void setLon(double lon) {this.lon = lon;} 
+	public void setRadius(double radius) {this.radius = radius;} 
 }
